Prawie działający build :D

+++ ProjektPP/rectangle.h
@@ -6,5 +6,5 @@ class rectangle : public shape
 	public:
 		rectangle(const point* start, const int* color);
 	
-		void draw(short int** img = 0);
+		void draw(int** img = 0);
 };
